requirements.txt
streamlit==1.31.1
opencv-python-headless
pillow==10.2.0
numpy==1.26.4pip